backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-autoslug==1.9.9
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
Markdown==3.10.1
sqlparse==0.5.5
tzdata==2025.3
